Run finalize scripts after installing kernel
I needed a way to adjust the GRUB config, but there are no hooks between
writing the bootloader config and building the disk image.

I considered adding a new hook for it, but I could not find any use of
the finalize hook in the wild that wouldn't work just as well at this
slightly later point in the build process.